requirements.txt
nltk==3.8.1
spacy==3.7.4
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
huggingface-hub==0.23.2
python-dotenv==1.0.1
transformers==4.41.1
pandas==2.2.2
sentence-transformers==2.7.0
evaluate==0.4.2
mauve-text==0.3.0
bert-score==0.3.13
sentencepiece==0.2.0
absl-py==2.1.0
rouge_score==0.1.2
groq==0.8.0
langchain-core==0.2.1
langchain-chroma==0.1.1
levenshtein==0.25.1
openai==1.30.4
rank-bm25==0.2.2
langchain-community==0.2.1
